chore: update config files
