--- .eslintignore
@@ -1,8 +1,0 @@
-.d2/
-node_modules/
-.git/
-.vite/
-dist/
-build/
-src/types/dhis2-openapi-schemas/
-src/locales/
