+++ .prettierrc.mjs
@@ -1,0 +1,1 @@
+export { default } from '@dhis2/config-prettier'
