+++ package.json
@@ -25,7 +25,10 @@
         "@cypress/vite-dev-server": "^7.2.1",
         "@dhis2/cli-app-scripts": "^12.11.0",
         "@dhis2/cli-style": "^10.7.8",
-        "@dhis2/cypress-commands": "^10.1.0",
+        "@dhis2/config-eslint": "^0.2.2",
+        "@dhis2/config-prettier": "^0.2.2",
+        "@dhis2/cypress-commands": "^10.1.2",
+        "@eslint/compat": "^2.0.4",
         "@ls-lint/ls-lint": "^1.11.0",
         "@sonar/scan": "^4.3.5",
         "@testing-library/dom": "^10.4.0",
